requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
asyncio==4.0.0
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.2
blinker==1.9.0
Brotli==1.1.0
build==1.3.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
choreographer==1.1.2
chromadb==1.2.0
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
coverage==7.11.0
cryptography==46.0.3
cssselect2==0.8.0
cssutils==2.11.1
curl_cffi==0.13.0
cycler==0.12.1
dash==3.2.0
dash-bootstrap-components==2.0.4
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
durationpy==0.10
email-validator==2.3.0
et_xmlfile==2.0.0
faiss-cpu==1.12.0
fastapi==0.119.1
filelock==3.20.0
filetype==1.2.0
financetoolkit==2.0.6
finmodels==2.0.4
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.9.0
google-api-core==2.26.0
google-auth==2.41.1
google-cloud-aiplatform==1.121.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.15.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.45.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
greenlet==3.2.4
grpc-google-iam-v1==0.14.3
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.35.3
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaleido==1.1.0
kiwisolver==1.4.9
kubernetes==34.1.0
langchain==1.0.1
langchain-anthropic==1.0.0
langchain-classic==1.0.0
langchain-community==0.4
langchain-core==1.0.0
langchain-google-genai==3.0.0
langchain-openai==1.0.0
langchain-text-splitters==1.0.0
langgraph==1.0.0
langgraph-checkpoint==2.1.2
langgraph-prebuilt==1.0.0
langgraph-sdk==0.2.9
langsmith==0.4.37
logistro==2.0.0
loguru==0.7.3
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
mdurl==0.1.2
mmh3==5.2.0
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.9.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.4
oauthlib==3.3.1
onnxruntime==1.23.1
openai==2.5.0
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
orjson==3.11.3
ormsgpack==1.11.0
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pdfplumber==0.11.7
peewee==3.18.2
pillow==11.3.0
platformdirs==4.5.0
plotly==6.3.1
pluggy==1.6.0
posthog==5.4.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.0
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycares==4.11.0
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.2.5
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.17.2
PyPika==0.48.9
pyproject_hooks==1.2.0
pyrate-limiter==3.9.0
pyreadline3==3.5.4
pytesseract==0.3.13
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
regex==2025.9.18
reportlab==4.4.4
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retrying==1.4.2
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
sec-downloader==0.11.2
sec-edgar-downloader==5.0.3
sec-parser==0.58.1
shapely==2.1.2
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
starlette==0.48.0
streamlit==1.50.0
sympy==1.14.0
tabulate==0.9.0
tavily-python==0.7.12
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.22.1
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typer==0.19.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.38.0
watchdog==6.0.0
watchfiles==1.1.1
weasyprint==66.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
win32_setctime==1.2.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
yfinance==0.2.66
zipp==3.23.0
zopfli==0.2.3.post1
zstandard==0.25.0
